fix floor-parser: rename service and extend wall tests
