+++ services/floor_parser/internal/entities/geometry.py
