+++ services/floor_parser/internal/entities/raw_plan.py
