+++ services/floor_parser/internal/export/floor_exporter.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
-from services.parser.internal.entities.floor import FloorPlan
+from services.floor_parser.internal.entities.floor import FloorPlan
+from services.floor_parser.internal.entities.warnings import ParseWarning
 
 
 class FloorExporter:
@@ -8,7 +9,8 @@ class FloorExporter:
         self,
         floor_plan: FloorPlan,
         source: str = "dxf",
-        units: str | None = None
+        units: str | None = None,
+        warnings: list[ParseWarning] | None = None,
     ) -> dict[str, object]:
         return {
             "schema_version": floor_plan.schema_version,
@@ -20,7 +22,8 @@ class FloorExporter:
             "walls": [self._export_wall(wall) for wall in floor_plan.walls],
             "doors": [],
             "windows": [],
-            "rooms": []
+            "rooms": [],
+            "warnings": [warning.to_dict() for warning in (warnings or [])],
         }
 
     def _export_wall(self, wall) -> dict[str, object]:
