+++ services/floor_parser/tests/apartment_outline_polyline.json
@@ -1,0 +1,112 @@
+{
+  "schema_version": "0.1.0",
+  "meta": {
+    "source": "dxf",
+    "source_ref": "apartment_outline_polyline.dxf",
+    "units": "mm"
+  },
+  "walls": [
+    {
+      "id": "8B:1",
+      "points": [
+        [
+          0.0,
+          0.0
+        ],
+        [
+          7000.0,
+          0.0
+        ]
+      ],
+      "width": 0.0
+    },
+    {
+      "id": "8B:2",
+      "points": [
+        [
+          7000.0,
+          0.0
+        ],
+        [
+          7000.0,
+          1800.0
+        ]
+      ],
+      "width": 0.0
+    },
+    {
+      "id": "8B:3",
+      "points": [
+        [
+          7000.0,
+          1800.0
+        ],
+        [
+          9000.0,
+          1800.0
+        ]
+      ],
+      "width": 0.0
+    },
+    {
+      "id": "8B:4",
+      "points": [
+        [
+          9000.0,
+          1800.0
+        ],
+        [
+          9000.0,
+          6000.0
+        ]
+      ],
+      "width": 0.0
+    },
+    {
+      "id": "8B:5",
+      "points": [
+        [
+          9000.0,
+          6000.0
+        ],
+        [
+          0.0,
+          6000.0
+        ]
+      ],
+      "width": 0.0
+    },
+    {
+      "id": "8B:closing",
+      "points": [
+        [
+          0.0,
+          6000.0
+        ],
+        [
+          0.0,
+          0.0
+        ]
+      ],
+      "width": 0.0
+    },
+    {
+      "id": "8C",
+      "points": [
+        [
+          3000.0,
+          0.0
+        ],
+        [
+          3000.0,
+          3200.0
+        ]
+      ],
+      "width": 0.0
+    }
+  ],
+  "doors": [],
+  "windows": [],
+  "rooms": [],
+  "warnings": []
+}
