+++ services/floor_parser/tests/apartment_partition_lines.json
@@ -1,0 +1,98 @@
+{
+  "schema_version": "0.1.0",
+  "meta": {
+    "source": "dxf",
+    "source_ref": "apartment_partition_lines.dxf",
+    "units": "mm"
+  },
+  "walls": [
+    {
+      "id": "8B",
+      "points": [
+        [
+          0.0,
+          0.0
+        ],
+        [
+          9000.0,
+          0.0
+        ]
+      ],
+      "width": 0.0
+    },
+    {
+      "id": "8C",
+      "points": [
+        [
+          9000.0,
+          0.0
+        ],
+        [
+          9000.0,
+          6000.0
+        ]
+      ],
+      "width": 0.0
+    },
+    {
+      "id": "8D",
+      "points": [
+        [
+          9000.0,
+          6000.0
+        ],
+        [
+          0.0,
+          6000.0
+        ]
+      ],
+      "width": 0.0
+    },
+    {
+      "id": "8E",
+      "points": [
+        [
+          0.0,
+          6000.0
+        ],
+        [
+          0.0,
+          0.0
+        ]
+      ],
+      "width": 0.0
+    },
+    {
+      "id": "8F",
+      "points": [
+        [
+          3500.0,
+          0.0
+        ],
+        [
+          3500.0,
+          6000.0
+        ]
+      ],
+      "width": 0.0
+    },
+    {
+      "id": "90",
+      "points": [
+        [
+          3500.0,
+          2500.0
+        ],
+        [
+          9000.0,
+          2500.0
+        ]
+      ],
+      "width": 0.0
+    }
+  ],
+  "doors": [],
+  "windows": [],
+  "rooms": [],
+  "warnings": []
+}
