+++ services/floor_parser/tests/square_room.json
@@ -1,0 +1,70 @@
+{
+  "schema_version": "0.1.0",
+  "meta": {
+    "source": "dxf",
+    "source_ref": "square_room.dxf",
+    "units": "mm"
+  },
+  "walls": [
+    {
+      "id": "8B",
+      "points": [
+        [
+          0.0,
+          0.0
+        ],
+        [
+          6000.0,
+          0.0
+        ]
+      ],
+      "width": 0.0
+    },
+    {
+      "id": "8C",
+      "points": [
+        [
+          6000.0,
+          0.0
+        ],
+        [
+          6000.0,
+          5000.0
+        ]
+      ],
+      "width": 0.0
+    },
+    {
+      "id": "8D",
+      "points": [
+        [
+          6000.0,
+          5000.0
+        ],
+        [
+          0.0,
+          5000.0
+        ]
+      ],
+      "width": 0.0
+    },
+    {
+      "id": "8E",
+      "points": [
+        [
+          0.0,
+          5000.0
+        ],
+        [
+          0.0,
+          0.0
+        ]
+      ],
+      "width": 0.0
+    }
+  ],
+  "doors": [],
+  "windows": [],
+  "rooms": [],
+  "warnings": []
+}
