+++ services/floor_parser/tests/test_parse_floor.py
@@ -4,19 +4,31 @@ import json
 import unittest
 from pathlib import Path
 
-from services.parser.internal.classification.classifier import SemanticClassifier
-from services.parser.internal.export.floor_exporter import FloorExporter
-from services.parser.internal.normalization.geometry_normalizer import GeometryNormalizer
-from services.parser.internal.readers.dxf.extractor import DxfExtractor
-from services.parser.internal.readers.dxf.reader import DxfReader
-from services.parser.internal.topology.topology_builder import TopologyBuilder
+from services.floor_parser.internal.classification.classifier import SemanticClassifier
+from services.floor_parser.internal.export.floor_exporter import FloorExporter
+from services.floor_parser.internal.normalization.geometry_normalizer import GeometryNormalizer
+from services.floor_parser.internal.readers.dxf.extractor import DxfExtractor
+from services.floor_parser.internal.readers.dxf.reader import DxfReader
+from services.floor_parser.internal.topology.topology_builder import TopologyBuilder
 
 
 class ParseFloorIntegrationTest(unittest.TestCase):
     def test_square_room(self):
+        self._assert_floor_json_matches_expected("square_room.dxf", "square_room.json")
+
+    def test_apartment_partition_lines(self):
+        self._assert_floor_json_matches_expected("apartment_partition_lines.dxf", "apartment_partition_lines.json")
+
+    def test_apartment_outline_polyline(self):
+        self._assert_floor_json_matches_expected("apartment_outline_polyline.dxf", "apartment_outline_polyline.json")
+
+    def test_floorplan(self):
+        self._assert_floor_json_matches_expected("floorplan.dxf", "floorplan.json")
+
+    def _assert_floor_json_matches_expected(self, dxf_filename: str, json_filename: str) -> None:
         tests_dir = Path(__file__).resolve().parent
-        dxf_path = tests_dir / "square_room.dxf"
-        expected_json_path = tests_dir / "square_room.json"
+        dxf_path = tests_dir / dxf_filename
+        expected_json_path = tests_dir / json_filename
 
         result = self._parse_dxf(dxf_path)
