--- services/parser/tests/square_room.json
@@ -1,45 +1,0 @@
-{
-  "schema_version": "0.1.0",
-  "meta": {
-    "source": "dxf",
-    "source_ref": "square_room.dxf",
-    "units": "mm"
-  },
-  "walls": [
-    {
-      "id": "2F",
-      "points": [
-        [0.0, 0.0],
-        [6000.0,0.0]
-      ],
-      "width": 0.0
-    },
-    {
-      "id": "30",
-      "points": [
-        [6000.0,0.0],
-        [6000.0, 5000.0]
-      ],
-      "width": 0.0
-    },
-    {
-      "id": "31",
-      "points": [
-        [6000.0, 5000.0],
-        [0.0, 5000.0]
-      ],
-      "width": 0.0
-    },
-    {
-      "id": "32",
-      "points": [
-        [0.0, 5000.0],
-        [0.0, 0.0]
-      ],
-      "width": 0.0
-    }
-  ],
-  "doors": [],
-  "windows": [],
-  "rooms": []
-}
